Create python time
Python function that is passed four integer inputs representing two different time periods (each time period is expressed as a number of minutes and seconds). The name of the function should be seconds_between() and it should calculate and return the time difference between those two time periods (expressed as a number of seconds).